Prepare prefixed trees using a temporary index file.
The current approach of running "git commit" in the user's worktree
pollutes the reflog, could leave an inconsistent state if "braid update"
is killed, and presents a gratuitous obstacle to adding a --no-commit
option that would allow pre-existing uncommitted changes.  Using a
temporary index file avoids all of these problems.

Add tests of "braid update" with mergeable and conflicting
project-specific changes (which pass on the old code and demonstrate
that the new code hasn't broken these cases) and a test that the reflog
isn't polluted (which fails on the old code).