Tests: Set the committer identity per-repo, not globally.
In the VM I was using for Braid development, my global identity is
intentionally unset so that I'm prompted to choose an identity for each
repository.  Inasmuch as this is a valid use case, the test suite should
not change developers' global configuration without their consent.

We may as well set the identity unconditionally as a simplification.