Include non-mirror local content in the base and remote trees again.
When I changed Braid to prepare trees using a temporary index file in
092b8c5, I skipped reading non-mirror local content into the base and
remote trees, thinking it was a harmless optimization.  But when git
diffs the base and local trees, this can cause it to falsely detect a
file deleted from the mirror as being renamed to some unrelated file in
the local tree.  With this change, the trees are constructed the same
way as before 092b8c5.

Add a test case and rename some of the existing fixtures so I can
remember what they're for.

Fixes #41.