Rename the configuration file from .braids to .braids.json.
Fixes #48.

Sometimes end users like to inspect and interact with the .braids.json file
directly and would like their editor to detect it as json. The easiest way to
achieve this is to add the .json extension. Existing braid setups are migrated
from the old configuraiton filename to the new one next time the configuration
file is modified in anyway. Typically this is when adding, removing or udpating
a mirror.