Default to removing the remotes from the repository after any operation except setup.
This default behaviour can be overridden by passing --keep command line options.

The motivations is that downstream users end up with large number of remotes (sometimes
as high as 20-30) that are difficult to work with and not really intended to be interacted
with.

There can also be some issues where the remotes are updated in .braids.json but
are not updated in repository resulting in some unexpected behaviour. Removing the
remotes avoids the scenario altogether.

Fixes #47