When a mirror has a nonempty remote path, use the basename of the remote path (instead of that of the repository URL) as the default local path.
Update the tests for mirrors of subdirectories to specify the mirror
path.  This means we lose the coverage of the default path, but it's
pretty low-impact if this breaks.  If we used the new default path of
"layouts" (the basename of the remote path, "layouts"), then we'd lose
the coverage of unequal local and remote paths, which is a much more
serious concern.