Set up type checking using Sorbet and annotate parts of the code.
See the new "Type checking" section of development.md for a full
explanation.

I've tried to make this change as low-risk as possible while still
achieving the goal of type checking.  Notably, I've done some tricks so
that Braid doesn't use or depend on the Sorbet runtime library at all
during normal use by end users, but when Braid is run as part of the
test suite, it uses the Sorbet runtime library and runtime type checks
to help catch bugs and annotation mistakes.

I want to keep this change focused on adding type checking, so I've kept
code changes other than the addition of type checking infrastructure and
type annotations to a minimum.  While working on type checking, I
discovered two new bugs in Braid, but I just added FIXME comments and am
not fixing the bugs in this change.  The fixes can be done in later pull
requests with appropriate test cases.