Require Braid code to be run with the correct gems already activated.
This ensures that the correct versions of Braid and its dependencies are
used, which is important now that we have a dependency (Sorbet) that
reserves the right to make breaking changes in any update.

This change should make no difference to end users using an installed
Braid gem.  However, Braid developers will now have to go through
Bundler when running the `braid` executable or commands such as rake or
rspec.  This is what we want because the behavior of the Braid test
suite may be affected by which version of sorbet-runtime it uses, for
example.

The `braid` launcher is moved from `bin/braid` to `exe/braid`, as is
standard for gems developed with Bundler.  This frees up the `bin/braid`
path for a binstub generated by `bundle binstubs braid`.  Braid
developers have to run `bundle binstubs braid` manually, but then they
can run `bin/braid` as before.  Alternatively, they can use `bundle exec
braid`.

The new check supersedes the previous ad-hoc code in `bin/braid` that
added Braid's `lib` directory to the load path but did not do anything
about Braid's dependencies.