Get most files in `lib` to `typed: true`.
The exceptions are check_gem.rb and version.rb, which can't safely
depend on the Sorbet runtime.

My motivation is to get the IDE extension to use type information in all
of these files.  In particular, a `typed` sigil of `true` or higher
appears to be needed for results from a file to be included in a search
for references.  Having complete search results is important to help us
avoid mistakes when working on the code.  It seems odd that the IDE
extension refuses to work on files with `typed: false` given that the
documentation (https://sorbet.org/docs/static) claims that Sorbet does
use type information from such files at least in some cases, but I don't
see any way around that restriction right now and it's not too hard to
get to `typed: true`.

Of course, the available type information is still limited by the lack
of annotations, but this is an important step forward.