Type-check most of operations.rb and fix handling of spaces.
(Since each of these changes requires extensive reworking of
operations.rb and review of all its callers, it's efficient to make both
changes together.)

To fix the handling of spaces (and perhaps other characters as well) in
shell command arguments, we just keep the arguments as a list of strings
all the way to an Open3.capture3 (or similar) that doesn't use the
shell, instead of joining the arguments with spaces.  I added one
integration test of each Braid command with spaces in the file paths.

I'm still mostly avoiding nontrivial changes to Braid's runtime
behavior, preferring to leave them to a separate pull request.  A
notable exception: when generating a remote name, Braid now replaces a
much larger set of characters with underscores.  For the new tests to
pass, I had to at least replace spaces, and I decided to go all the way.

Fixes #88.